+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.19.0 (2024-xx-xx)
 
+- Improvement: better formatting for voteban results
 - Other: internal improvements
 
 ## 1.18.0 (2024-02-29)